+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6.7
 
 * fix(`TraceData`): Do not throw exception on missing data.
+* fix(`MainTransformer`): Improve error reporting on circular references.
 
 ## 0.6.6